// bookscape-ui/src/app/services/models/book-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookDto {
  archived?: boolean;
  authorName?: string;
  cover?: Array<string>;
  createdBy?: number;
  createdDate?: string;
  id?: number;
  isbn?: string;
  lastModifiedBy?: number;
  lastModifiedDate?: string;
  owner?: string;
  rate?: number;
  shareable?: boolean;
  synopsis?: string;
  title?: string;
}